Reworked the meatatheory until the definitions of atomic actions.
Now, we need to decide how to change the send-wrappers in the presence of semantic hooks.
At the moment it seems that we need to fish out all of the hooks from the world used to
define the action and then show frameability with respect to the irrelevant hooks or the hooks
that are weaker than those that were used to define the action.

Exciting stuff!